requirements.txt
h5py
numpy
pandas
pillow
gdown
plotly
einops
webdataset
matplotlib
torch
torchvision
lifelines
scikit-learn
umap-learn
umap-learn[plot]
scikit-survival
tqdm
seaborn==0.12.0
hydra-core==1.2.0
opencv-python==4.4.0.46
patchify